move views in __init__ to separate file
